feat(conf): do not override extensions if exist

+++ src/config/system/userActivationScripts/vscode/activate.sh
@@ -1,12 +1,12 @@
 echo 'settings vscode...'
 
-# Reset everything
-rm -rf /data/vscode
-
 # Create --user-data-dir
+rm -rf /data/vscode/data
 mkdir -p /data/vscode/data/User
 echo "${argSettings}" > /data/vscode/data/User/settings.json
 
 # Create --extensions-dir
-mkdir -p /data/vscode/extensions
-cp -L --no-preserve=mode -R "${argExtensions}/share/vscode/extensions/"* /data/vscode/extensions
+if ! test -e /data/vscode/extensions; then
+  mkdir -p /data/vscode/extensions
+  cp -L --no-preserve=mode -R "${argExtensions}/share/vscode/extensions/"* /data/vscode/extensions
+fi
